hydra/code/requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
netaddr==0.7.19
pySecurityCenter==3.0.3
requests==2.18.4
urllib3==1.22
